// src/torrent/WebTorrentTypes.d.ts
    interface TorrentDestroyOptions {
        destroyStore?: boolean | undefined;
    }
// ...
    interface Torrent extends NodeJS.EventEmitter {
        readonly infoHash: string;

        readonly magnetURI: string;

        readonly torrentFile: Buffer;

        readonly torrentFileBlobURL: string;

        readonly files: TorrentFile[];

        readonly announce: string[];

        readonly ['announce-list']: string[][];

        readonly pieces: Array<TorrentPiece | null>;

        readonly timeRemaining: number;

        readonly received: number;

        readonly downloaded: number;

        readonly uploaded: number;

        readonly downloadSpeed: number;

        readonly uploadSpeed: number;

        readonly progress: number;

        readonly ratio: number;

        readonly length: number;

        readonly pieceLength: number;

        readonly lastPieceLength: number;

        readonly numPeers: number;

        readonly path: string;

        readonly ready: boolean;

        readonly paused: boolean;

        readonly done: boolean;

        readonly name: string;

        readonly created: Date;

        readonly createdBy: string;

        readonly comment: string;

        readonly maxWebConns: number;

        destroy(opts?: TorrentDestroyOptions, cb?: (err: Error | string) => void): void;

        addPeer(peer: string | SimplePeer): boolean;

        addWebSeed(url: string): void;

        removePeer(peer: string | SimplePeer): void;

        select(start: number, end: number, priority?: number, notify?: () => void): void;

        deselect(start: number, end: number, priority: number): void;

        createServer(opts?: RequestOptions): Server;

        pause(): void;

        resume(): void;

        on(event: 'infoHash' | 'metadata' | 'ready' | 'done', callback: () => void): this;

        on(event: 'warning' | 'error', callback: (err: Error | string) => void): this;

        on(event: 'download' | 'upload', callback: (bytes: number) => void): this;

        on(event: 'wire', callback: (wire: Wire, addr?: string) => void): this;

        on(event: 'noPeers', callback: (announceType: 'tracker' | 'dht') => void): this;
    }

    interface TorrentFile extends NodeJS.EventEmitter {
        readonly name: string;

        readonly path: string;

        readonly length: number;

        readonly downloaded: number;

        readonly progress: number;

        select(): void;

        deselect(): void;

        createReadStream(opts?: { start: number, end: number, }): NodeJS.ReadableStream;

        getBuffer(callback: (err: string | Error | undefined, buffer?: Buffer) => void): void;

        appendTo(
            rootElement: HTMLElement | string,
            opts?: { autoplay?: boolean | undefined, controls?: boolean | undefined, maxBlobLength?: number | undefined },
            callback?: (err: Error | undefined, element: HTMLMediaElement) => void): void;
        appendTo(rootElement: HTMLElement | string, callback?: (err: Error | undefined, element: HTMLMediaElement) => void): void;

        renderTo(
            rootElement: HTMLMediaElement | string,
            opts?: { autoplay?: boolean | undefined, controls?: boolean | undefined, maxBlobLength?: number | undefined },
            callback?: (err: Error | undefined, element: HTMLMediaElement) => void): void;
        renderTo(rootElement: HTMLMediaElement | string, callback?: (err: Error | undefined, element: HTMLMediaElement) => void): void;

        getBlob(callback: (err: string | Error | undefined, blob?: Blob) => void): void;

        getBlobURL(callback: (err: string | Error | undefined, blobURL?: string) => void): void;
    }

    interface TorrentPiece {
        readonly length: number;

        readonly missing: number;
    }